Delegate to selected property components.
The ids are rertieved from the layout where the current component
is found. When getting the completions for a newly added property
we failed to use the actual components of the selected property.
Add a delegate to the selected property.

Fixes: 146678111
Test: Unit test added.
Change-Id: I1f4dfc1a76cfae7513340305fc7228e40ee80579
(cherry picked from commit b16c91d6713e2a1aa191d121bc73d741db39e470)