[Dsl] preserve invariant in MultiType property models
PropertyTransforms assume that if transform() is called, the
transform's .test() passes.  This is true most of the time, but for a
period while MultiType properties are being updated (through
setTypeAndValue()), the element and the transform might be out of
sync, while the new element is generated and bound.  For that period,
return null from getElement().

Bug: n/a
Test: existing
Change-Id: Ifa6b9d7005c3de2c1ce98bbaf90965ca67b0ce66