[Dsl] remove previous implementation of Set type
Now that the model's property description is authoritative for model
elements of type set, we can remove the previous implementation of
passing a boolean through from model to parser and back.

We do not attempt to detect errors by the Dsl writer here, so there's
no special support for detecting e.g. a listOf expression for a Set
property; this means that we only need to deal with the property's
type on write.

Bug: 141842964
Bug: 148657110
Test: existing
Change-Id: Icbb01790e391b914959d3ec98657023e82ee97bb