Remove BulkFileListener from DatabaseInspectorProjectService.
BulkFileListener does not work as we assumed.
BulkFileListener::before is called before the VFS is updated,
but after the file is deleted from disk.
When this was implemented in the Database Inspector we believed
BulkFileListener::before to be called before the file was deleted.
Since we had a wrong assumption BulkFileListener is used incorrectly
here. Furthermore it is causing a test to be flaky.

Test: NA
Bug: 148576272
Change-Id: Ife8aeafadc7f89cbbbd2b587e4ac7f92e5c56160