Fix cancellation in ListenableFuture<I>.finallySync
Previous implementation of finallySync didn't follow guava's contract
of deep cancellation. Futures.transform docs describes this contract as
following:
The returned future attempts to keep its cancellation state in sync with
that of the input future. That is, if the returned future is cancelled,
it will attempt to cancel the input, and if the input is cancelled,
the returned ruture will receive a callback in which it will attempt
to cancel itself.

Previously:
1. Input wasn't cancelled when returned future was cancelled, because it
didn't listen for cancellation on returned future;
2. If input was cancelled then returned was completed as "failed" instead of
"cancelled", because it didn't check if received exception is
CancellationException and simply called setException().

bug: 151713370
Test: FutureUtilsTest.testFinallySyncCancelled
Change-Id: I79ecd5a8c8728caa5f987047bd61b6c0c8ed18a7