Don't change combobox selection if one exists already
We should not let the combobox selection be changed (by the service) if a process is already selected. Let user select a difference process manually.

This also alleviates startup performance penalty of opening the tool window on a device or emulator that has many debuggable apps open, because it will not try to attach to every one of them.

Test: new AppInspectionProcessComboModelTest
Fixes: 151377288
Change-Id: I40bf8cde2986a626a877111d81d33b649750058f