Change empty state message.
The current empty state message is misleading: "select a process to
begin inspecting". What is actually happening is that the db inspector
is waiting for the app to start a connection to the database.
This CL changes the message accordingly.

This change is temporary, since eventually the Database Inspector
tab will be always visible by default, we will have to introduce
two states:
First state: app not started, message: "Start your app to being inspecting"
Second state: after the app starts but before a connection to the
database is started, message: "Waiting for the app to open a connection
to the database..."

Test: NA
Bug: 148049132
Change-Id: I541513b8187acd93d700c8f2b7aac219e5f2581d