Close all tabs when last db is closed.
So far when a database was closed we only removed tabs associated
with it.
After this CL this remains the same, except when the last database
is removed, in which case we remove its tab + all the SqliteEvaluator
tabs. They should not be open since there is no database to run queries
against.

Test: DatabaseInspectorControllerTest.kt
Bug: 148049132
Change-Id: I329532410f3a016690b16052e2866347909381b8