Fix table rows diff bug after alter table.
If a table with number of rows > 0 is opened from the schema panel
and an "alter table add column" statement is run on that table,
the rows diff is made against the old list of rows.
This doesn't work well becuase the new list of rows has a
different number of columns.
This CL fixes this bug by deleting the old list of rows when the
columns change. This causes the diffing to recreate all the rows
in the UI.

Test: TableControllerTest.kt
Bug: 151079468
Change-Id: I54fee23a47732670a15001c71a44c000d7dba3e8