Refactoring: Move ExpandableLabel out as separate component
This functionality was added to CollapsibleLabelPanel and
InspectorPanelImpl. Extract the functionality and simplify.
This gives us a Label component which can be expanded with a popup
to the right of the text if there is not enough horizontal space
for the text.
This work is done in preparation for fixing b/151910312.

Bug: 151910312
Test: Existing tests should pass
Change-Id: Ib74d1ae50fd620e860049f6603a40d5e8bf0c6fe