Fix Connection Assistant stateful button
When the Connection Assistant is opened, the StatefulButton starts
scanning for connected USB devices. Once it has the results after a few
seconds, it updates the panel. But the text is too long without
wrapping, so the entire panel gets scrolled all the way to the right.
This change makes it fit in the original panel size.

Bug: 152075176
Test: manual
Change-Id: I3d861beb78600ae70311d53c30077fc0ef773504