[Dsl] DependenciesInfoModel file definitions
Use the fact that we now have an interface to implement to split out a
small number of test file definitions from the large -Impl enum, and
instead place the definitions locally to the test code.

In terms of line count, this is not actually an improvement, but the
locality of the definitions and the reduction in redundant
information (no need now to have name prefixes or to repeat the
directory each time) is on balance a simplification.

Bug: n/a
Test: existing (no change in functionality)
Change-Id: I0a7ad79405ebfe781cd00883f67e92f08d5eda7e