[Dsl,.gr] when deleting, delete slightly more
The code for deleting dsl elements in Groovy was theoretically
insufficient, in that elements represented by method calls end up with
PsiElements for the Dsl element (used for modifications) being
substantially different from the overall Psi needed to represent the
element as a whole; for example

  signingConfigs {
    foo {
      storePassword file('abc.txt')
    }
  }

would represent the STORE_PASSWORD property with a GradleDslMethodCall
whose PsiElement corresponded just to the `file('abc.txt')` portion.

Deletion of this kind of structure tended to work, but only because
almost all elements have isInsignificantIfEmpty() turned on, and those
that don't (buildTypes / productFlavors) turn out not to have elements
commonly represented in this form.  However, an upcoming change to
signingConfig, to treat empty elements as significant (since even
mentioning the signing config creates it) revealed this problem.  When
deleting, continue deleting parents until a suitable parent is
reached.

Bug: 150196538
Test: added SigningConfig test
Change-Id: I4ce1d3afbae5a736d786c67bd43479dd89f3ef70