Make NavXmlIndex failure handling capture all errors
Unfortunately, it looks like the JAXB parsing logic throws more than
just UnmarshalException, as I just saw a case where it throws an
AssertionError.

Due to the lack of the API's ability to follow its own contract, we have
to treat it with uncertainty. And since we definitely don't want
any exception leaking to users here, we just catch all possible
exceptions aggressively.

Bug: 150487407
Test: Covered by existing tests
Change-Id: If2f36880edb4eff2b6deafe1cde2d8f235df2c00