[Sync Error Handlers] Convert more error handlers.
In this change :
Delete Java HeapSpaceErrorHandler: This errorHandler offers links on how
to increase the vn options for Gradle in case these are very low. This
information is already mentioned in gradle-properties file in new
versions of Android Studio. The default size is now 2040MB.
Conevrt InternetConnectionErrorHandler to
InternetConnectionIssueChecker.
Convert MissingAndroidPluginErrorHandler to
MissingAndroidPluginIssueChecker : Delay the checking of google maven
repository existence to the time when the quickFix is being executed.
Fetch for the Android plugin in build files as Gradle models can't be
trusted when sync fails.
Convert MissingAndroidSdkErrorHanlder to MissingAndroidSdkIssueChecker :
Offer opening the file in the quickFix without pointing to the line that
has sdk.dir property.

Bug: 150060053
Bug: 149822714
Test: added
Change-Id: I8bd266631006f6b8afeefcc471afcdec487a78aa