Add a banner control on top of the layout inspector
When using compatibility mode for an API 29 or above device display
add a warning in the banner.

Update the "Live Updates" checkbox to be disabled when the current
client is not a live inspector client. [This was enabled when the
LegacyClient was used for a API 29 device.]

Also for the new tests:
- set the currentClient in LayoutInspector
- use the featureApiLevel when creating a the iDevice
- when attaching a LegacyDevice use the stream/process instances
  created by LegacyDevice (otherwise it will not connect correctly)

Fixes: 152349406
Test: Added unit test for banner and updated other unit tests.
Change-Id: I02f5d41ca8ec958597af7c491c8c0df94e4289f9