Rename UI-code references from Atrace to SystemTrace
Atrace currently means Atrace or Perfetto and is used in the UI to
specify a type of profilling. As we move to make the distinction between
Atrace and Perfetto more clear on the backend, this change will help
avoiding any confusion in betweent UI and backend representations across
our code-base.

Bug: 147099951
Test: n/a
Change-Id: I6942f502705ea900f05a89e1f41c2dbbd233c922