Disable view correction feature on dumb mode
The RenderErrorContributor has a feature to suggest similar views when a
view is not found and similar classes exist.
This requires to run in smart mode and, although the method should not
be called in dumb mode, a number of crashes have showed up.
This is a minimal fix that disables that feature when called in dumb
mode.
Also, try to avoid calling the render error contributor when in dumb
mode. Just a second line of defense.

Fixes: 149357583
Test: New regression test passes
Change-Id: Iae0e591d23ca76eaaced6a4f327a61aef01fb5ff