Update App Inspection look and feel for single inspector
For its initial release, we're only going to launch a single inspector
in the app inspection tool window. Therefore, this CL reskins it so it
looks as if it is just a database inspector, rather than a collection
of multiple inspectors.

To do this, we...

1) rename the tool window to "Database Inspector"
2) hide the tabbed pane so users don't see it
3) update the icon to look like a database tool window

We also take advantage of this CL to show a "There are no selected
processes" error message.

Fixes: 150865123
Fixes: 150386063
Bug: 152556591
Test: Updated
Change-Id: Iab7e80cdd6d5d17756b075daa2c245f1b0d0aafe