Make RenderTask#measureChild to be async
This moves the responsability of times out up the chain to Scene. Scene
can now decide how much it wants to wait for the measurment.
This is just a workaround to avoid blocking. The only drawback is that,
worst case scenario, the component resizing will not snap to
match_parent correctly.

This also adds timeouts for testing when calling measureChild so
deadlocks do not cause timeouts which are hard to debug.

Bug: 150159800
Test: Updated tests pass
Change-Id: I5b7bdc582a6592acf31cf7f1ce6e4fc508b2261f