Allow tests loading projects to be lenient on SyncIssues.
AGP is now more strict about presence of a manifest during
sync and will output an error if it's not there.

While we could fix the test projects to include this file,
it's also quite useful to have the infrastructure for tests
to be able to ignore specific SyncIssue when the tests does
not care.

This change introduces this ability and makes the tests
that don't care about missing manifest use it.

Bug: N/A
Test: N/A
Change-Id: Iae0f082d194d4e87eb96841d4f202fef0d086a9b