Adds a support for qualifiers for Dagger editing.
Qualifier is an annotation that has a `javax.inject.Qualifier` annotation.

A Dagger's binding is defined not only by a type, but also by a qualifier.
Return provider for consumer and vise verse only if they have the equal qualifiers or if both don't have any.
Qualifiers are equal when all their attribute's values are equal.

Fixes: 152481333
Test: DaggerUtilTest
Change-Id: I9c33d406ecfffc83b470d3240872180a58fd724d