[Dsl] allow getting the untransformed element for a model
If a model has a non-trivial transform, then the element returned by
GradlePropertyModelImpl.getElement() is not in the same logical place
in the Dsl as the model element.  Using the transformed element, which
could be arbitrarly deeply nested in the Dsl, would give spurious
dereferencing in the fully qualified name.

Bug: 151216354
Test: new SigningConfigModel test in upcoming CL
Change-Id: Iafab66374ea93a82f25462c03f473c7287087cae