[Dsl,.kts] also rename references to subproperties
Although the code here to perform the rename operation looks mostly
generic, it does not in fact work for Groovy, for reasons related to
b/152603582.  It is nevertheless useful to have a test case for
KotlinScript so that we can continue working on handling the edge
cases of renaming while having a safety net.

Bug: 147867235
Test: added SigningConfigModel test
Change-Id: I4cda4ae1b9897a9c2c5063194be60b063b8068ad