Remove maximum width for validity on store dialog
There was a limit of 60 pixels on the width of spinner for the
validity years of just 60 pixels, causing this component to not
always be visible.

This CL removes the maximum width and also sets the initial number of
columns of the text field of the spinner to 2.

Bug: 151765078
Test: Manually confirmed that size was changed.
Change-Id: I7b87f6a797d0cfc1c55f526f706d7e025532bf2b
(cherry picked from commit 9e3ed14349f80400e7b176b49a58527ba82c91de)