SettingFragment not added from menu
When SettingsFragment was added from the add destination menu
a new destination was not created on the design surface.
PreferenceFragmentCompat is not in the default dependencies so
the type information is not avaible from NavigationSchema
until the background sync completes.

Default to fragment as the tag name if the type information
is not available. This should work unless a fragment with a
custom tag and a custom dependency is added to the new fragment
wizard gallery.

Fixes: 150035127
Test: testCreateSettingFragment
Change-Id: I58fe6b43b77ef9a682e56fee571a372502dba4c5