[Dsl] distinguish between external- and internal-syntax references
Most of the time, when resolving references we are resolving text that
has been written in an external syntax, and therefore needs to be
converted to our internal representation of hierarchical model names.
However, sometimes, notably when we construct a ReferenceTo a model
object ourselves, we do not have an external syntax, but instead write
directly in our internal form.  We must not at that point try to
convert from an external syntax again, otherwise we might trip
exceptions or return nonsense.

Bug: 151216354
Bug: 152603582
Test: existing (more coming up)
Change-Id: I4cb34ebc3fceb075ee1c7f7fae7a6be9f40bf0d4