[Dsl] one less dubious split-on-dots operation
Since elements (such as signing configs) can have dots in their names,
we need to be careful not to treat such dots as hierarchy dots.  We've
already done most of the work for this, in being careful with
GradleNameElements; we just need to use the appropriate helper
routines when converting a string form of a GradleName to a list
indicating hierarchy.

Bug: 126937269
Bug: 134739625
Bug: 151216354
Test: upcoming
Change-Id: Id0e7ffdf4007518908bed5cbcb5c73a7f6e5b3e5