[Dsl] provide a means for hierarchical renaming
We need an escape hatch for the KotlinScript-specific way of referring
to extra properties in array syntax, ext["prop"], as `ext.prop` as
opposed to just `prop`, and we need that while not prohibiting
renaming models to properties with dots in their name.  However, this
probably doesn't do what a user might expect it to do, and as such
should probably not be exposed in the long term; in particular,
renaming things to hierarchical names in other places in the model
hierarchy is unlikely to work.

Bug: 148769031
Bug: 151216354
Bug: 152603582
Test: adjusted existing test
Change-Id: Ia595e565ce2be9f8ca7d276690219c4bbc6871c6