[Dsl] maybeUpdateName needs to unescape
When writing out the result of a rename operation, the name that
maybeUpdateName receives comes from the GradleNameElement, which is in
principle an escaped name (dots and backslashes handled separately.
This means we need to unescape the name before further treatment by
the language-specific writer.  Define an unescape() operation, and use
it at the appropriate point in maybeUpdateName.

Also update existing tests to use """ ... """ Kotlin string literals
where backslashes are involved, to perhaps lessen confusion over what
is a real backslash.

Bug: 148769031
Bug: 151216354
Bug: 152603582
Test: add escape/unescape tests
Change-Id: Idf33cf5fd2d460af46c184840fafeccd13160966