Initial support for Preview data sources
This adds the parsing code for PreviewParameters within Previews. This
will generate ParametrizedPreviewElements that can generate multiple
PreviewElement instances.
Each of this instances will represent a unique preview in the UI and in
the deploy to device.

This CL does not contain the code to "instantiate"
ParametrizedPreviewElements into single PreviewElement instances. This
will be in a follow-up CL.

Bug: 150783335
Test: New tests for parsing pass
Change-Id: I23944c1fd939fb6ee98fc8e5eceb66d9904a1d51