[Sync Error Handlers] Convert more Error handlers.
convert FailedToParseSdkErrorHandler to FailedToParseSdkIssueChecker.
Convert MissingPlatformErrorHandler to MissingPlatformIssueChecker.

These errorHandlers need to use the project instance. For now the
solution is to loop over all the opened projects and their modules: as
long as we come to the module which path is the same as the issueData
projectPath, then we found the module/project needed.

Bug: 149822714
Test: added
Change-Id: Ie4c488b50b6a706141dc0cf3980c18c6e98280fd