Changes installation default to be only on current user
On API 24 or newer by default the option --user current
is added at installation time. This removes the confusion
where if the app is already installed in not the current
user by default it's updated only there causing run to
fail after that as we run on the current user.

Fixes: 69491108
Bug: 135376049
Fixes: 133772367
Fixes: 139706026
Bug: 136509779
Test: Updated unit tests
Change-Id: Ice8ea39ecb2122928d4b115781e76c14afdb0a88