[Gradle Sync] Setup models via startup instead of sync
This change adjusts AndroidGradleProjectStartupActivity to instead of
syncing (via a possible cached sync) to just attach all models in the
cache to their respective facets. The result of this is that the build
tool window is not opened nor are any of the sync events triggered.

This allows us to simplify sync by removing the cached sync and brings
us closer to being able to completely rely on intellijs refreshProject.

Bug: N/A
Test: manually tested
Change-Id: Ibd05172c749fc84be4cdfc45b97dc9956ad47de9