Wait for sync to complete when clicking finish in some NPW fixtures
The way we wait for sync to complete will change and we prepare the
codebase for these changes by avoiding direct waiting for sync to
complete.

This way we also add waiting in some places where we didn't wait for
sync to complete. In general UI tests should not abandon running sync
as it is not fully supported by the test framework right now, thus
this change, even if it may make several tests a little slower, should
improve overall stability.

Bug: n/a
Test: n/a
Change-Id: I192b94846f5423e5ff2a6cec9ce0d2fa00b6a130