Fix Lint tooltip when error message contains a url
Sometime in 2018 JetBrains decided to change the UX for rendering
inspection tooltips. One of the changes was to strip away the
"more..." link that many inspection messages had, and instead convert
that link into a dedicated "Show Inspection Info" action.
(See b/153084296 for more info.)

To accomplish that, DaemonTooltipWithActionRenderer searches for the
first HTML href in the inspection message and assumes it is the
inspection description link. Of course, that heuristic breaks if the
inspection message contains an href of its own.

It's not clear whether this is a bug in IDEA, or whether web urls in
inspections messages are unsupported. But we can work around it fairly
easily by moving the inspection description link to the beginning
of the inspection message. After all, it is not user-visible anymore
anyway.

Fixes: 153164864
Test: manual---I could not find a reasonable way to test the behavior of
      DaemonTooltipWithActionRenderer in a unit test.
Change-Id: I1109c6a457a059c601749589626123ee642e1e74