Fix potential race condition in BaseAction update
There seems to be some faulty bookkeeping in XDebugSession
and its associated RunContentDescriptor. This manifests in
BaseAction's update where it tries to query an active debug
session without an associated window.

I was not able to reproduce the issue, but I hope this CL
will work around the issue on JetBrain's end.

Change-Id: I73552aab8f2310890185595f96e2b02c8b6dc1f8
Fixes: 149344306
Test: N/A