Do NOT throw a status exception when no windows are found.
When attaching to a running process on a pre 29 device we used to
throw an exception when no windows could be found. When the
layout inspector was trying to auto-connect we would retry later.

However the same code would throw the same exception when the user
would try to attach to a running process on the device.
Change the logic to simply return a boolean instead.

Fixes: 149581576
Test: Added unit test. (one extra unit test on master)
Change-Id: I405fe6f27901c6cb144310c243f86dae9e6319fa