Support adding of service event listeners in AppInspectorClient
Currently, we have no way of knowing whether a client has been disposed.
This CL allows us to add listeners to listen in on crash and dispose
events.

This CL also refactors AppInspectorClient.EventListener by moving its onCrashEvent and onDispose functions to a new interface named ServiceEventListener, and allows the adding of said ServiceEventListener to AppInspectorClient. The motivation behind this is to support multiple listeners for crash/dispose events. For example: a listener is needed by the inspection library to know when a crash occurs and show a toast.

Test: new EventNotifierTest and updated existing tests.
Bug: 153383819
Change-Id: I3df94e89751a98252adb18168bd9219bb748f279