Refactor the launch inspector aspect of discovery service
The frontend does not need to know about AppInspectionTarget. All It needs to know are the processes that it can attach to, as well as the information related to launching an inspector (ex: inspectorId and jar). This CL removes the part of discovery service that allows listening to new AppInspectionTarget, and changes the function AppInspectionView calls into discovery. Instead of calling attachToProces to obtain an AppInspectionTarget, AppInspectionView can directly launch an inspector via launchInspector. Discovery service will take care of the caching of targets and clients. As a result, discovery is now manager of AppInspectorClient.

Test: updated tests.
Bug: n/a
Change-Id: Ifd362edea4d57718e4c6c92855e3963a35844ee5