Support variant switching from cached variants
If build files are not changed when switching build variants, then
save the current variant models so that they can be used without
Gradle Sync if it is switched back again.

This change saves the current variant models to project user data,
then retrieve them when setting up AndroidModuleModel/NdkModuleModel,
and inject the saved variants to the new model.

Bug: 152889824
Test: BuildVariantUpdaterIntegTest
Change-Id: Ifc6f82506afbea4fa6e485dc14d6c3b433feb02d