Import UTP results
This CL implements an Android Studio action to import UTP results and
display them in the new multi device test result panel.

The action will open up a window to let the user to select a (binary)
protobuf test result file. Then it will be loaded and displayed in the
new test result panel.

This CL does not have error handling. In fact, we would need a UX mock
first before we can implement error messages with GUI.

Bug: 152773288
Test: ImportUtpResultActionTest, UtpTestResultAdapterTest
Change-Id: I489a8726a0d825e2c9b59ef692d097700b0676e1