Fix LegacyClientTest
The problem was that we weren't always setting
InspectorClient.clientFactory back correctly (even more so after
I1379f0d7), and so sometimes when we came here there wasn't a
DefaultInspectorClient to find.
I1379f0d7 made it possible to not create a DefaultInspectorClient in the
first place, so doing that, and also fixing other places where the
factory is changed.

Test: this is a test fix
Fixes: 150701560
Change-Id: I7c1d4f4aca0cf11c74ec789d65e6840037463d9e