Support cleaning up of clients based on project
This CL allows the front end (AppInspectionView) to clean up all of the clients it has started, without touching clients started by other projects.

Projects also automatically attempt to dispose all of their clients when they are in the process of shutting down. Therefore, under normal circumstances, studio will be able to reattach to a running process without needing to restart it.

Test: new AppInspectionDiscovery disposal test.
Fixes: 153114370
Fixes: 151765060
Change-Id: I53e413ed08508675b0b71cd3e37b172b3e243a93