Include generated sources in GradleTestArtifactSearchScopes
This includes the paths for generated sources so the files coming from
annotation processors like Dagger can be correctly identified. This is
important for the scope enlarger to provide the correct scope and not
showing incorrectly unresolved references in the editor.
Usually, this sources would be included by AGP in the test artifacts
but it will not until the next version.

Fixes: 151029089
Test: Adds regression test and existing tests pass
Change-Id: I2076e1d452ad8151a63b1fa2ddece3ce947c41e4
(cherry picked from commit 4c75d6a6710dc412bda916dfda737098552849be)