Fix TransportEventListener not being removed even if requested
The existing code is running `TransportEventListener` and concurrently
updating a flag indicating if the listener should be removed after.
Unfortunately, most of the time, the check on this flag happens before
it is updated concurrently.
As a consequence, the callback from `MemoryProfilerStage` is often
needlessly retained, which keeps the profiler stage live even after
the user has left it. If the stage happens to be holding a heap dump,
this consumes much more memory than neccessary.

The new code behaves slightly differently.
In the original code, the last listener execution decides whether
the listener should be removed. In the fixed code, if any call decides
the listener should be removed, it is removed. Studio-wise, this
should cause no change, because all existing callbacks have been
returning constants.

Bug: 154041292
Bug: 152344964
Test: existing, manual inspection through VisualVM
Change-Id: I4b0d80e1aac030b76ad564e6e673f16a05567972