Refactor DatabaseConnection implementations to use SqliteStatementType
By checking SqliteStatementType we now know if we're running an update
statement or a query statement. As a consequence is now possible to be
precise when using DatabaseConnection.
This CL splits the old `execute` method into two: `query` and `execute`.

1. `query` doesn't run any statement, it is just used to
get a SqliteResultSet.
The SqliteResultSet can be used to access data on the database.
To support EXPLAIN statements SqliteResultSet has been split
in two classes: `PagedSqliteResultSet` and `LazySqliteResultSet`.
By doing this we do paging differently based on the statement's
type. Using limit-offset for SELECT statements and locally
for EXPLAIN statements. This is necessary because to do
limit-offset paging we need to wrap the original statement
in a new statement, this is not possible with EXPLAIN statements.

2. `execute` is used to execute a statements, the result of
the statement is ignored. Therefore this method is ideal
to run data manipulation statements such as UPDATE, INSERT etc.

Test: intellij.android.debuggers.tests
Bug: 153684332

Change-Id: I26a349bb994f159de660e4302124b35a229a5967