[Gradle Sync] Remove Gradle wrapper clean up
This step ensures that a wrapper is present and if not creates one. This
is already handled by intellij, however there are some differences. The
intellij one creates the wrapper based off the Gradle settings page
which can be configured in the setting dialog. The intellij one also
creates a wrapper the the version used within the tooling api where as
ours creatres the latest supported one.

Bug: N/A
Test: N/A
Change-Id: I03590e4a076827c4e594eef2d95202b6a6f2aa96