Remove spammy ADB path mismatch warning.
AndroidSdkUtils#getAdb will log a warning every time it is called if the
adb location specified by android.adb.path doesn't point to a valid
file. The method is called a lot so it adds a lot of spam to userlogs.
This CL removes this warning because:
1. It's only used internally and not publicly documented anywhere.
2. Internally filed bugs attach system properties.  We can check that
   instead of relying on this warning anyway.

Bug: 154242118
Test: This CL removes a warning statement. No logic changes.
Change-Id: I0c86b1e1b1d332958652372d0bc603b92965914b
(cherry picked from commit 40f3a9bbbae5830d1c4fd8546e6342629905c761)