Move IdeInfo to android-common
IdeInfo is static information and there is no need to mock it. Better
alternatives are (1) move unnecessary plugins from Android plugin to
Android Studio plugin or (2) override the value in test specific
constructor. Note, that simply replacing the service in integration
tests is dangerous as the test still runs with the wrong set of plugins.

Any tests that does not make sense in non-Android Studio environment
should not run instead of making `.isAndroidStudio()` return `true`
without changing the set of applied plugins.

Bug: 147586218
Test: n/a
Change-Id: I717ad305b6fba72980ec8d1fb331557bc9f093a9