[sync,import] commit document when updating wrapper properties
GradleWrapper writes properties files using a low-level API, which
automatically updates a datestamp (and therefore modifies the file).
In order that high-level views of the file (e.g. Psi) are consistent,
we need to ensure that the Project's document corresponding to this
file is committed.

This means that where we have a Project and are calling a
`GradleWrapper` factory, we need to pass that Project through in order
that the Project can be used to obtain the PsiFile and Document
corresponding to the properties file.  Alter the factories to have an
explicit, nullable Project argument; at each call site, pass a Project
if there is one available.

In cases where we do not have a Project, for example because the
creation of a Gradle wrapper is part of a step of importing a project,
the existing behaviour (not committing the document) is safe; it is
only an issue if we are performing this low-level file manipulation
within the context of an existing Project.  Callers of
GradleWrapper.create() and friends in this situation are updated to
pass an explicit `null` to the factories.

Fixes: 154243928
Test: existing, should fail less often
Change-Id: I19ee2010dc173e40923c4d7e0bdc93da24ae7ee8