Reduce logging level from error to warn
The purpose of logging in AppInspectionView was to surface errors that appeared in async calls. However, because logging level was set to error, it caused our presubmit to fail tests even for expected exceptions such as future cancellation. Therefore, this CL reduces logging level to warn and also ignores FutureCancellationException altogether.

Test: n/a
Fixes: 154700214
Change-Id: I91669ffd05e30b95bef50140f94e2f6b7dfb11fa