[Dsl,.kts] sort references to extra["foo"] properties better
The existing logic to sort extra properties in order that later
properties should refer to values of earlier ones (and not vice versa)
compared external injection syntax (in arbitrary build language) with
GradleNameElement syntax, which generally worked OK with Groovy and
failed with Kotlin extra properties expressed as map dereferences.
Convert the property name to an internal syntax before checking
whether it is referenced in other properties.

It's not clear to me that the logic is correct even after this change:
I think that disguised references through projects will not be picked
up.  But this change should bring the two currently-supported build
languages to parity in this respect.

Bug: 138277630
Test: added
Change-Id: I32c6f1f7b773d644099cac5cad01d3350987de06