Disable run button when sqlite statement has parsing errors
Since we use the parser to determine the type of a SQLite statement
and then make assumptions on the type, the tool might behave
unexpectedly if we can't determine the type of the statement
because of parsing errors.
By preventing the user to run statements that we can't recognize
we avoid this problem.

Test: SqliteEvaluatorViewImplTest.kt, SqliteEvaluatorControllerTest.kt
Bug: 154895984
Change-Id: I2116591e27dc1b7eb620d285188b6e998a54b467