[Dsl] complete audit of test cases
Subclasses of GradleFileModelTestCase, which currently run
parameterized as Groovy or Kotlin, now have isKotlinScript to mirror
isGroovy; at present they are complements of each other, but in
principle a third language could come along.

Annotate the currently-ignored tests with information about whether
they are ignored because they are:

- irrelevant in one or other language;
- exposing a bug in the support for one or other language (but passing
  in the other).

If they are exposing a bug, ensure that there is a TODO comment with a
reference to an issue in the bugtracker nearby, in a format that can
be picked up by todo-tracking tools such as Squeegee.

Bug: 138277630
Test: existing
Change-Id: I81c2aca1e3f3459f6ff47623ea2e5cecd98bf842