Display heap dump in separate stage, but not live allocation
This CL starts a new stage for displaying the heap dump.
It also tags the `DurationData` class with whether the
(yet to be loaded) capture will be heap-dump-like in order
to be displayed in this separate stage.

The profiler combobox is not yet working properly in this
new heap dump stage, which will be fixed in a separate CL.

Bug: 154270916
Bug: 149595558
Test: move existing, ignore obsolete
Change-Id: If3e985eac61fb8b485db3768c368f7f464595581