[Dsl] fail tests when test files are missing
This effectively changes the default in Dsl tests, when a test case
file has not been submitted to the repository, from an implicit
pass (the effect of ignore) to an explicit fail.  The intent behind
this change is to ensure that new tests are at least actively
considered for all supported build languages, not just one.

All existing tests now having been considered and -- where applicable
-- having had an explicit explanation for why they should be ignored,
this completes the audit of our existing test files.

Fixes: 138277630
Test: existing
Change-Id: If1c1ad19cb3b0f486bec8c193e45a4145117d59a