Quickfix to enable desugaring
... when build fails when a project uses Java 7 but there are
dependencies with Java 8.

This change adds a couple of quickfix that sets java level to 8 (if it
is lower) and also minSdkVersion to 26 or 24 depending on the error. The
quickfixes apply tha changes to the module that failed or to all
modules.

The quickfix uses a processor to show the places that will change and
the tooltip on those usages describe what the change will do.

Bug: 154759255
Test: DexDisabledIssueCheckerTest, EnableDexWithApiLevelQuickFixTest
Change-Id: Ic89929460d6a4a1c5007506502ce3cdc2c17a6f0