Add home/parent stage properties, implement for heapdump stage
This CL allows each stage to customize its own "parent" stage
to go back to and "home" stage to reset to.

It also implements that behavior for the heap dump stage:
either clicking back button or re-selecting the "MEMORY"
item goes back to the memory profiler stage.

The menu re-selection behavior matches that of the CPU Capture stage,
which was an accident turned into a requested feature.

Bug: 149595558
Test: add new
Change-Id: I9bf4d4053a8dddabea367619bb12d6802427fcc4