[Dsl,.kts] add a complex property syntax test
This is perhaps a showcase for the potential for ugliness in the
KotlinScript Dsl: a lot of things which are conceptually simple
dereferences need type decorations, explicit project prefixes, and
other noise.  Of course the noise means that the name of injections is
substantially different from the Groovy case, and not very nice to
encode, so only check the name in Groovy.

One of the tests, renaming map entries which should not resolve to
each other but currently do, is not well-founded; elide those test
parts when testing the KotlinScript parser.

Bug: 138277630
Bug: 149286032
Test: added
Change-Id: Iaa993acf360fbc3d215dd377b34b037f4e48c444